Tweaked and tuned and hacked all around.
Here's the thing, though - my current hacks adds a m_var to glscene,
which tells it where to shift the mesh.

Now, of course, if the object we're controlling isn't the last in
stack, this change propagates onwards. fixable bt resetting the
m_var at the end of scene drawing, but still...


git-svn-id: https://jeshmup.googlecode.com/svn/trunk@77 abe21748-ea51-0410-af0e-c35c4496dcea